add method to list cursos on controller

+++ src/CursoOnline.Dominio/Cursos/CursoParaListagemDto.cs
@@ -6,5 +6,5 @@ public class CursoParaListagemDto
     public string Nome { get; set; }
     public double CargaHoraria { get; set; }
     public string PublicoAlvo { get; set; }
-    public double Valor { get; set; }
+    public decimal Valor { get; set; }
 }
